refactor(insights): extract palettes, move defaultCostColumns, tighten Column type, drop dead i18n
