+++ frontend/packages/web/components/admin/insights/cost/palettes.ts
@@ -1,0 +1,28 @@
+export const PALETTE_WORKSPACE = [
+  '#1e40af',
+  '#1d4ed8',
+  '#2563eb',
+  '#3b82f6',
+  '#60a5fa',
+  '#93c5fd',
+  '#bfdbfe',
+]
+export const PALETTE_MODEL = [
+  '#166534',
+  '#15803d',
+  '#16a34a',
+  '#22c55e',
+  '#4ade80',
+  '#86efac',
+  '#bbf7d0',
+]
+export const PALETTE_USER = [
+  '#6b21a8',
+  '#7e22ce',
+  '#9333ea',
+  '#a855f7',
+  '#c084fc',
+  '#d8b4fe',
+  '#e9d5ff',
+]
+export const PALETTE_CACHE = ['#b45309', '#d97706', '#f59e0b', '#fbbf24', '#fcd34d']
